[TASK] Use later hook for replacing mailto:-Links.
The previous hook was too early, non-cached parts of the page were not yet inserted, thus the extension didn't work when an email address is displayed in the template of an uncached plugin.

This probably has a small performance drawback, but the regex search was run on each page view anyway. It just has to parse additional content (the uncached content).